Update simulation_functions.R
Allows for a reference group to be specified. If a reference group is specified, the model is first estimated in the simulated data sets for the reference group to generate parameter estimates for the items specific to the reference group. The parameter estimates for the linking items, derived from group 1, are then constrained when estimating the model in the full sample. This sets the metric of the scale to the mean of the reference group, which facilitates group mean comparisons. This update also allows for group sample sizes to be specified individually, allowing for unbalanced groups.